+++ src/middlewares/index.js
@@ -1,16 +1,25 @@
-import checkUserAnonymous from "./checkUserAnonymous";
-import checkUserExist from "./checkUserExist";
-import checkUserLoggedIn from "./checkUserLoggedIn";
-import errorHandler from "./errorHandler";
-import imageUploader from "./imageUploader";
-import logPathHistory from "./logPathHistory";
-import saveResponseLocalData from "./saveResponseLocalData";
-import checkItemExist from "./checkItemExist";
-import checkItemOwner from "./checkItemOwner";
-import checkPostExist from "./checkPostExist";
-import checkPostOwner from "./checkPostOwner";
-import checkCommentExist from "./checkCommentExist";
-import checkCommentOwner from "./checkCommentOwner";
+// common
+import errorHandler from "./common/errorHandler";
+import imageUploader from "./common/imageUploader";
+import logPathHistory from "./common/logPathHistory";
+import saveResponseLocalData from "./common/saveResponseLocalData";
+
+// comment
+import checkCommentExist from "./comment/checkCommentExist";
+import checkCommentOwner from "./comment/checkCommentOwner";
+
+// item
+import checkItemExist from "./item/checkItemExist";
+import checkItemOwner from "./item/checkItemOwner";
+
+// post
+import checkPostExist from "./post/checkPostExist";
+import checkPostOwner from "./post/checkPostOwner";
+
+// user
+import checkUserAnonymous from "./user/checkUserAnonymous";
+import checkUserExist from "./user/checkUserExist";
+import checkUserLoggedIn from "./user/checkUserLoggedIn";
 
 export {
   checkUserAnonymous,
